# android/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "picweight API",
    "description": "AI-assisted calorie & БЖУ tracker. Photograph what you eat; a server-side estimation agent returns calories and macros.",
    "license": {
      "name": "MIT"
    },
    "version": "master"
  },
  "paths": {
...
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Refresh the session JWT",
        "description": "Accepts a currently-valid session (bearer header or cookie) and issues a new JWT with a fresh expiry. An expired session cannot be resurrected — that path is a full re-login. Cookie callers also get the cookie re-set.",
        "operationId": "refresh",
        "responses": {
          "200": {
            "description": "A new session JWT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Session missing, invalid or expired",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
    "/api/v1/meals/events": {
      "get": {
        "tags": [
          "meals"
        ],
        "summary": "Stream meal analysis events",
        "description": "Server-sent events for this user only. Each `completed`, `reanalyzed` and `group_settled` event carries the full day state so a client can render the notification without a follow-up request.",
        "operationId": "meal_events",
        "responses": {
          "200": {
            "description": "text/event-stream; each frame is a JSON MealEvent under the `meal` event name",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/MealEvent"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DayState": {
        "type": "object",
        "description": "A day's totals against its targets.\n\nCarried verbatim in the `/meals/events` completion payload so a client\nrenders the notification from one message with no follow-up round trip (§9).",
        "required": [
          "date",
          "consumed_kcal",
          "target_kcal",
          "remaining_kcal",
          "consumed_protein_g",
          "target_protein_g",
          "remaining_protein_g",
          "consumed_fat_g",
          "target_fat_g",
          "consumed_carbs_g",
          "target_carbs_g",
          "meals_logged",
          "status"
        ],
        "properties": {
          "consumed_carbs_g": {
            "type": "number",
            "format": "double",
            "description": "Carbohydrate consumed, grams."
          },
          "consumed_fat_g": {
            "type": "number",
            "format": "double",
            "description": "Fat consumed, grams."
          },
          "consumed_kcal": {
            "type": "number",
            "format": "double",
            "description": "Energy consumed, kcal."
          },
          "consumed_protein_g": {
            "type": "number",
            "format": "double",
            "description": "Protein consumed, grams."
          },
          "date": {
            "type": "string",
            "format": "date",
            "description": "The local calendar day these figures cover."
          },
          "meals_logged": {
            "type": "integer",
            "format": "int64",
            "description": "Meals logged on this day (all statuses except `failed`)."
          },
          "remaining_kcal": {
            "type": "number",
            "format": "double",
            "description": "`target_kcal - consumed_kcal`; may be negative."
          },
          "remaining_protein_g": {
            "type": "number",
            "format": "double",
            "description": "Protein still owed, grams; clamped at zero."
          },
          "status": {
            "$ref": "#/components/schemas/DayStatus",
            "description": "Classification derived from the figures above."
          },
          "target_carbs_g": {
            "type": "number",
            "format": "double",
            "description": "Carbohydrate allowance, grams."
          },
          "target_fat_g": {
            "type": "number",
            "format": "double",
            "description": "Fat floor, grams."
          },
          "target_kcal": {
            "type": "number",
            "format": "double",
            "description": "Energy target, kcal. Zero when targets are not computed."
          },
          "target_protein_g": {
            "type": "number",
            "format": "double",
            "description": "Protein floor, grams."
          }
        }
      },
      "DayStatus": {
        "type": "string",
        "description": "How a day is going. Computed from the numbers, never asked of a model.",
        "enum": [
          "on_track",
          "tight",
          "over",
          "protein_unreachable",
          "no_targets"
        ]
      },
      "ErrorBody": {
        "type": "object",
        "description": "The JSON body returned for every [`AppError`].",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "Stable machine-readable code, e.g. `not_found`."
          },
          "message": {
            "type": "string",
            "description": "Human-readable detail. Generic for 5xx, specific for 4xx."
          }
        }
      },
...
      "MacroTotals": {
        "type": "object",
        "description": "Summed macros, used for meals, days and groups alike.",
        "required": [
          "kcal",
          "protein_g",
          "fat_g",
          "carbs_g"
        ],
        "properties": {
          "carbs_g": {
            "type": "number",
            "format": "double",
            "description": "Carbohydrate, grams."
          },
          "fat_g": {
            "type": "number",
            "format": "double",
            "description": "Fat, grams."
          },
          "kcal": {
            "type": "number",
            "format": "double",
            "description": "Energy, kcal."
          },
          "protein_g": {
            "type": "number",
            "format": "double",
            "description": "Protein, grams."
          }
        }
      },
...
      "MealEvent": {
        "type": "object",
        "description": "One event on the stream.\n\n`Clone` because broadcast hands every subscriber its own copy.",
        "required": [
          "kind",
          "revision",
          "emitted_at"
        ],
        "properties": {
          "body": {
            "type": [
              "string",
              "null"
            ],
            "description": "Line 2 + 3 + 4 of the notification, already phrased."
          },
          "day": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DayState",
                "description": "The day state, so no follow-up request is needed."
              }
            ]
          },
          "dish_name": {
            "type": [
              "string",
              "null"
            ],
            "description": "Dish name — the notification leads with it (§6)."
          },
          "emitted_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the event was produced, UTC."
          },
          "error": {
            "type": [
              "string",
              "null"
            ],
            "description": "Failure reason, set only on `Failed`."
          },
          "group_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "The sitting, when the meal belongs to one."
          },
          "headline": {
            "type": [
              "string",
              "null"
            ],
            "description": "Ready-to-render notification lines."
          },
          "kind": {
            "$ref": "#/components/schemas/MealEventKind",
            "description": "What happened."
          },
          "meal_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "The meal, for everything except `GroupSettled`."
          },
          "revision": {
            "type": "integer",
            "format": "int32",
            "description": "Revision these figures describe."
          },
          "status": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/MealStatus",
                "description": "Meal status after the event."
              }
            ]
          },
          "totals": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/MacroTotals",
                "description": "Totals for the meal, or for the whole sitting on `GroupSettled`."
              }
            ]
          }
        }
      },
      "MealEventKind": {
        "type": "string",
        "description": "What happened.",
        "enum": [
          "queued",
          "analyzing",
          "completed",
          "reanalyzed",
          "failed",
          "group_settled",
          "updated"
        ]
      },
...
      "MealStatus": {
        "type": "string",
        "description": "Lifecycle of a meal. `pending`/`analyzing` are in-flight, `needs_review`\nmeans the agent produced a draft the user has not confirmed, `confirmed`\nis the only status `recall_similar_meals` reads from, `failed` is loud.",
        "enum": [
          "Pending",
          "Analyzing",
          "NeedsReview",
          "Confirmed",
          "Failed"
        ]
      },
...
      "TokenResponse": {
        "type": "object",
        "description": "A minted picweight session JWT.",
        "required": [
          "token",
          "expires_in"
        ],
        "properties": {
          "expires_in": {
            "type": "integer",
            "format": "int64",
            "description": "Lifetime in seconds.",
            "minimum": 0
          },
          "token": {
            "type": "string",
            "description": "The session JWT, for `Authorization: Bearer`."
          }
        }
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Session JWT. Web clients receive it as the HttpOnly `picweight_session` cookie after OIDC login; native clients obtain it from `POST /api/auth/token` and send it as `Authorization: Bearer`."
      }
    }
  },
...
}